removed separated component, migrated to shared one
