+++ src/modules/feature-modules/assessment/assessment.module.ts
@@ -13,9 +13,6 @@ import { SidebarInnovationMenuOutletComponent } from './base/sidebar-innovation-
 // Pages.
 // // Account.
 import { PageAssessmentAccountManageAccountInfoComponent } from './pages/account/manage-account-info.component';
-import { AssessmentPageAccountEmailNotificationsListComponent } from './pages/account/email-notifications/email-notifications-list.component';
-import { AssessmentPageAccountEmailNotificationsEditComponent } from './pages/account/email-notifications/email-notifications-edit.component';
-
 // // Dashboard.
 import { DashboardComponent } from './pages/dashboard/dashboard.component';
 // // Innovation.
@@ -43,8 +40,6 @@ import { AssessmentService } from './services/assessment.service';
     // Pages.
     // // Account.
     PageAssessmentAccountManageAccountInfoComponent,
-    AssessmentPageAccountEmailNotificationsListComponent,
-    AssessmentPageAccountEmailNotificationsEditComponent,
     // // Dashboard.
     DashboardComponent,
     // // Innovation.
