+++ src/modules/shared/pages/account/email-notifications/email-notifications-edit.component copy.html
@@ -1,0 +1,14 @@
+{% comment %} <theme-content-wrapper [status]="pageStatus">
+
+  <div class="nhsuk-grid-row">
+    <div class="nhsuk-grid-column-two-thirds">
+
+      <form [formGroup]="form">
+        <theme-form-radio-group controlName="notificationPreference" [label]="notificationPreferences.question" [items]="notificationPreferences.items"></theme-form-radio-group>
+        <button class="nhsuk-button nhsuk-u-margin-top-3" (click)="onSubmit()">Change preferences</button>
+      </form>
+
+    </div>
+  </div>
+
+</theme-content-wrapper> {% endcomment %}
