+++ src/modules/shared/pages/account/email-notifications/email-notifications-edit.component.html
@@ -1,14 +1,17 @@
 <theme-content-wrapper [status]="pageStatus">
+  <form [formGroup]="form" class="nhsuk-grid-column-two-thirds nhsuk-u-padding-left-0">
+      
+      <div class="nhsuk-warning-callout nhsuk-u-margin-top-6">
+          <div class="nhsuk-heading-m nhsuk-warning-callout__label">
+              <span role="text"> <span class="nhsuk-u-visually-hidden">Important: </span>You could miss time sensitive service reminders</span>
+          </div>
+          <p *ngIf="isAssessmentType">If you do not select ‘innovator submits innovation record’ or ‘assigned needs assessor’ notifications, you could miss time sensitive service reminders. </p>
+          <p *ngIf="isQualifyingAccessorRole">If you do not select ‘suggestions to support’, you could miss time sensitive service reminders.</p>
+      </div>
 
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-two-thirds">
-
-      <form [formGroup]="form">
-        <theme-form-radio-group controlName="notificationPreference" [label]="notificationPreferences.question" [items]="notificationPreferences.items"></theme-form-radio-group>
-        <button class="nhsuk-button nhsuk-u-margin-top-3" (click)="onSubmit()">Change preferences</button>
-      </form>
-
-    </div>
-  </div>
-
+      <theme-form-checkbox-array class="nhsuk-u-margin-bottom-4" arrayName="preferencesEnabled" [items]="formPreferencesList" [pageUniqueField]="false"></theme-form-checkbox-array>
+      
+      <button class="nhsuk-button nhsuk-u-margin-top-3" (click)="onSubmit()">Confirm</button>
+      
+  </form>
 </theme-content-wrapper>
